Introduces 'MultiModule'
MulitModule is a module container that holds other modules. Now modules
can be arbitrarily cascaded. This is useful to group modules.

Here it is used to group multiple ParticleGuns, in order to have one
configuration file for all particles solving user story 8.

It is also used to group the display modules.